+++ src/components/Blog/BlogPortableText/BlogCta.tsx
@@ -1,6 +1,6 @@
 import { PortableTextComponentProps } from "@portabletext/react";
 
-import { CTA } from "../../../node-lib/cms";
+import { CTA } from "../../../common-lib/cms-types";
 import { getCTAHref } from "../../../utils/portableText/resolveInternalHref";
 import ButtonAsLink from "../../Button/ButtonAsLink";