+++ src/components/pages/BlogIndex.page.tsx
@@ -9,10 +9,11 @@ import { useTheme } from "styled-components";
 
 import { BlogListJsonLd } from "../../browser-lib/seo/getJsonLd";
 import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
-import CMSClient, {
+import CMSClient from "../../node-lib/cms";
+import {
   BlogPostPreview,
   BlogWebinarCategory,
-} from "../../node-lib/cms";
+} from "../../common-lib/cms-types";
 import { decorateWithIsr } from "../../node-lib/isr";
 import BlogCategoryList from "../Blog/BlogCategoryList";
 import useBlogCategoryList from "../Blog/BlogCategoryList/useBlogCategoryList";