+++ src/utils/portableText/resolveInternalHref.ts
@@ -1,5 +1,4 @@
-import { CTA } from "../../node-lib/cms";
-import { CTAInternalLinkEntry } from "../../node-lib/cms/sanity-client/schemas";
+import { CTA, CTAInternalLinkEntry } from "../../common-lib/cms-types";
 import { assertUnreachable } from "../assertUnreachable";
 
 export const resolveInternalHref = (entry: CTAInternalLinkEntry): string => {